Refactor pystardog tests
