+++ docker-compose.cluster.yml
@@ -81,29 +81,12 @@ services:
       args:
         - TAG=${HAPROXY_TAG}
     container_name: ${STARDOG_LB}
+    ports:
+      - "127.0.0.1:5820:5820"
     depends_on:
       - zoo1
       - sd1
       - sd2
-
-  tests:
-    build:
-      context: .
-      dockerfile: dockerfiles/dockerfile-python
-    entrypoint: /bin/bash -c
-    command: ["./utils/run_tests.sh"]
-    container_name: pystardog_tests
-    environment:
-      - SSH_USER=${SSH_USER}
-      - SSH_PASS=${SSH_PASS}
-      - STARDOG_HOSTNAME_NODE_1=${STARDOG_HOSTNAME_NODE_1}
-      - STARDOG_HOSTNAME_CACHE=${STARDOG_HOSTNAME_CACHE}
-      - STARDOG_HOSTNAME_STANDBY=${STARDOG_HOSTNAME_STANDBY}
-      - STARDOG_ENDPOINT=http://${STARDOG_LB}:5820
-      - STARDOG_LB=${STARDOG_LB}
-      - STARDOG_USER=${STARDOG_USER}
-      - STARDOG_PASS=${STARDOG_PASS}
-
 
   # two mysql servers are used instead of one so we can simulate multiple datasources.
   mysql-music:
