--- dockerfiles/dockerfile-python
@@ -1,11 +1,0 @@
-FROM python
-
-RUN apt-get update && apt-get install sshpass jq -y
-
-COPY . /var/pystardog
-
-WORKDIR /var/pystardog
-RUN pip install -r requirements.txt -r test-requirements.txt
-
-
-
