+++ test/conftest.py
@@ -1,7 +1,9 @@
 import pytest
+import stardog
 import stardog.content as content
 import stardog.content_types as content_types
 import os
+
 
 # STARDOG_ENDPOINT = os.environ.get('STARDOG_ENDPOINT', None)
 STARDOG_HOSTNAME_NODE_1 = os.environ.get("STARDOG_HOSTNAME_NODE_1", None)
@@ -47,6 +49,7 @@ def ssl_verify(pytestconfig):
     return pytestconfig.getoption("ssl_verify")
 
 
+# this is currently not being used, need to confirm which one is definitive.
 @pytest.fixture
 def bulkload_content():
     contents = [
@@ -100,13 +103,174 @@ def music_options():
     return options
 
 
+@pytest.fixture()
+def admin(conn_string):
+    with stardog.admin.Admin(**conn_string) as admin:
+        yield admin
+
+
 @pytest.fixture
-def videos_options():
-    options = {
-        "jdbc.driver": "com.mysql.jdbc.Driver",
-        "jdbc.username": "user",
-        "jdbc.password": "pass",
-        "mappings.syntax": "STARDOG",
-        "jdbc.url": "jdbc:mysql://pystardog_mysql_videos/videos?useSSL=false",
-    }
-    return options
+def conn(conn_string, request):
+    endpoint = request.node.get_closest_marker("conn_endpoint", None)
+    username = request.node.get_closest_marker("conn_username", None)
+    password = request.node.get_closest_marker("conn_password", None)
+
+    if endpoint is not None:
+        endpoint = endpoint.args[0]
+        conn_string["endpoint"] = endpoint
+
+    if username is not None:
+        username = username.args[0]
+        conn_string["username"] = username
+
+    if password is not None:
+        password = password.args[0]
+        conn_string[password] = password
+
+    dbname = request.node.get_closest_marker("conn_dbname", None)
+    if dbname is not None:
+        with stardog.connection.Connection(dbname.args[0], **conn_string) as c:
+            yield c
+    else:
+        raise Exception("A dbname must be passed with the marker: conn_dbname")
+
+
+# The generic datasource is the music datasource.
+@pytest.fixture()
+def datasource(admin, music_options, request):
+    dsname = request.node.get_closest_marker("ds_name", None)
+    options = request.node.get_closest_marker("options", None)
+    if dsname is not None:
+        dsname = dsname.args[0]
+    else:
+        dsname = "datasource"
+
+    if options is not None:
+        options = options.args[0]
+    else:
+        options = music_options
+    ds = admin.new_datasource(dsname, options)
+    yield ds
+    ds.delete()
+
+
+@pytest.fixture()
+def virtual_graph(admin, music_options, datasource, request):
+    vgname = request.node.get_closest_marker("vgname", None)
+    mappings = request.node.get_closest_marker("mappings", None)
+    virtual_graph_options = request.node.get_closest_marker(
+        "virtual_graph_options", None
+    )
+    datasource_name = request.node.get_closest_marker("datasource_name", None)
+    database_name = request.node.get_closest_marker("database_name", None)
+    use_music_datasource = request.node.get_closest_marker("use_music_datasource", None)
+
+    if vgname is not None:
+        vgname = vgname.args[0]
+    else:
+        vgname = "virtual_graph"
+
+    if mappings is not None:
+        mappings = mappings.args[0]
+
+    if virtual_graph_options is not None:
+        virtual_graph_options = virtual_graph_options.args[0]
+
+    if database_name is not None:
+        database_name = database_name.args[0]
+
+    if use_music_datasource:
+        datasource_name = datasource.name
+
+    if virtual_graph_options is None and use_music_datasource is None:
+        raise Exception(
+            "Must pass either a datasource or VG options. See pytest marks in the pystest documentation"
+        )
+
+    vg = admin.new_virtual_graph(
+        vgname,
+        mappings=mappings,
+        options=virtual_graph_options,
+        datasource=datasource_name,
+        db=database_name,
+    )
+    yield vg
+    vg.delete()
+
+
+@pytest.fixture()
+def user(admin, request):
+    username = request.node.get_closest_marker("user_username", None)
+    password = request.node.get_closest_marker("user_password", None)
+    if username is not None:
+        username = username.args[0]
+    else:
+        username = "pystardogUser"
+
+    if password is not None:
+        password = password.args[0]
+    else:
+        password = "pystardogPass"
+
+    user = admin.new_user(username, password, False)
+    yield user
+    user.delete()
+
+
+@pytest.fixture()
+def role(admin):
+    role = admin.new_role("writer")
+    yield role
+    role.delete()
+
+
+@pytest.fixture
+def db(admin, request):
+    dbname = request.node.get_closest_marker("dbname", None)
+    options = request.node.get_closest_marker("options", None)
+    contents = request.node.get_closest_marker("contents", None)
+    kwargs = request.node.get_closest_marker("kwargs", None)
+
+    if dbname is not None:
+        dbname = dbname.args[0]
+    else:
+        dbname = "pystardog-test-db"
+
+    if options is not None:
+        options = options.args[0]
+
+    if contents is not None:
+        contents = contents.args[0]
+
+    if kwargs is not None:
+        kwargs = kwargs.args[0]
+
+    if contents is None and kwargs is None:
+        db = admin.new_database(dbname, options)
+    if kwargs is None and contents is not None:
+        db = admin.new_database(dbname, options, *contents)
+    if kwargs is not None and contents is None:
+        db = admin.new_database(dbname, options, **kwargs)
+    if kwargs is not None and contents is not None:
+        db = admin.new_database(dbname, options, *contents, **kwargs)
+    yield db
+    db.drop()
+
+
+@pytest.fixture
+def stored_query(admin, request):
+    query_name = request.node.get_closest_marker("query_name", None)
+    query = request.node.get_closest_marker("query", None)
+
+    if query_name is not None:
+        query_name = query_name.args[0]
+    else:
+        query_name = "everything"
+
+    if query is not None:
+        query = query.args[0]
+    else:
+        query = "select * where { ?s ?p ?o . }"
+    stored_query = admin.new_stored_query(query_name, query)
+    yield stored_query
+    stored_query.delete()
