+++ test/pytest.ini
@@ -1,0 +1,29 @@
+[pytest]
+markers =
+    # stored queries
+    query_name: name of a stored query
+
+    # databases
+    dbname: name of a database to be created
+    options: db options
+    contents: contents to pass to the db
+    kwargs: allow extra options to database creation. (only used to add the copy_to_server flag)
+
+
+    # datasources:
+    ds_name: name of the datasource to be created
+    options: options of the datasource.
+
+    # user:
+    user_username: name of the username to be create
+    user_password: password for the user to create.
+
+    # VG
+    vgname: name of the virtual graph to be created
+    use_music_datasource: Whether we want to use the music datasource as default
+    virtual_graph_options: virtual graph options
+    mappings: mappings to be used in a VG creation
+    database_name: name of the DB to associate the VG with.
+
+    # Connection
+    conn_dbname: database name to be used with the stardog client.
